refactor: 도메인 패키지 역할별 그룹화
