+++ README.md
@@ -38,4 +38,4 @@
 
 - [x] 단일 프로퍼티를 가지는 객체를 value class 사용으로 성능 최적화
 
-- [ ] 패키지를 역할별로 응집도 있게 재 조정
+- [x] 패키지를 역할별로 응집도 있게 역할별로 그룹화
