+++ src/main/kotlin/lotto/domain/model/winning/GainLoss.kt
@@ -1,4 +1,4 @@
-package lotto.domain.model
+package lotto.domain.model.winning
 
 enum class GainLoss {
     GAIN,
