+++ src/main/kotlin/lotto/domain/model/winning/WinTicketInfo.kt
@@ -1,5 +1,6 @@
-package lotto.domain.model
+package lotto.domain.model.winning
 
+import lotto.domain.model.lottoticket.LottoTicket
 import lotto.domain.valueobject.LottoNumber
 
 class WinTicketInfo(
