+++ src/test/kotlin/lottotest/domain/model/lottoticket/AutoLottoTicketTest.kt
@@ -1,7 +1,7 @@
-package lottotest.domain.model
+package lottotest.domain.model.lottoticket
 
-import lotto.domain.model.AutoLottoTicket
-import lotto.domain.model.LottoTicket
+import lotto.domain.model.lottoticket.AutoLottoTicket
+import lotto.domain.model.lottoticket.LottoTicket
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
