+++ src/test/kotlin/lottotest/domain/model/lottoticket/LottoTicketTest.kt
@@ -1,6 +1,6 @@
-package lottotest.domain.model
+package lottotest.domain.model.lottoticket
 
-import lotto.domain.model.LottoTicket
+import lotto.domain.model.lottoticket.LottoTicket
 import lotto.domain.valueobject.LottoNumber
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
