+++ src/test/kotlin/lottotest/domain/model/winning/RankTest.kt
@@ -1,6 +1,6 @@
-package lottotest.domain.model
+package lottotest.domain.model.winning
 
-import lotto.domain.model.Rank
+import lotto.domain.model.winning.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
